+++ plugin/BUILD
@@ -14,7 +14,7 @@ filegroup(
         "//plugin/pkg/admission/admit:all-srcs",
         "//plugin/pkg/admission/alwayspullimages:all-srcs",
         "//plugin/pkg/admission/antiaffinity:all-srcs",
-        "//plugin/pkg/admission/defaultpodnetworknotready:all-srcs",
+        "//plugin/pkg/admission/defaultpodnetworkreadiness:all-srcs",
         "//plugin/pkg/admission/defaulttolerationseconds:all-srcs",
         "//plugin/pkg/admission/deny:all-srcs",
         "//plugin/pkg/admission/eventratelimit:all-srcs",
